Make EvaluationJob independent from QuickRExView
